update elements on subscription page
