+++ components/admin_console/billing/billing_subscriptions/billing_subscriptions.scss
@@ -11,6 +11,10 @@
     }
 }
 
+#BillingSubscriptions__seatOverageTooltip {
+    margin-left: -5px;
+}
+
 .BillingSubscriptions__tooltip.tooltip {
     font-family: 'Open Sans', sans-serif;
