+++ components/admin_console/billing/plan_details/feature_list.scss
@@ -1,0 +1,20 @@
+.PlanDetailsFeature {
+    margin-top: 12px;
+    color: var(--sys-center-channel-color);
+    line-height: 16px;
+
+    i {
+        font-size: 14.4px;
+
+        &::before {
+            margin: 0;
+        }
+    }
+
+    span {
+        margin-left: 8px;
+        font-size: 12px;
+    }
+}
+
+
