+++ components/admin_console/billing/plan_details/index.tsx
@@ -5,48 +5,41 @@ import React from 'react';
 import {FormattedMessage} from 'react-intl';
 import {useSelector} from 'react-redux';
 
-import {getCurrentLocale} from 'selectors/i18n';
 import {GlobalState} from 'types/store';
 
+import {getCloudSubscription, getSubscriptionProduct} from 'mattermost-redux/selectors/entities/cloud';
+
 import {
-    seatsAndSubscriptionDates,
-    getPlanPricing,
     planDetailsTopElements,
     currentPlanText,
-    FeatureList,
 } from './plan_details';
+import FeatureList from './feature_list';
+import PlanPricing from './plan_pricing';
 
 import './plan_details.scss';
 
-type PlanDetailsProps = {
+type Props = {
     isFreeTrial: boolean;
     subscriptionPlan: string | undefined;
 }
-const PlanDetails = ({isFreeTrial, subscriptionPlan}: PlanDetailsProps) => {
-    const locale = useSelector((state: GlobalState) => getCurrentLocale(state));
+const PlanDetails = ({isFreeTrial, subscriptionPlan}: Props) => {
     const userCount = useSelector((state: GlobalState) => state.entities.admin.analytics!.TOTAL_USERS) as number;
-    const subscription = useSelector((state: GlobalState) => state.entities.cloud.subscription);
-    const product = useSelector((state: GlobalState) => {
-        if (state.entities.cloud.products && subscription) {
-            return state.entities.cloud.products[subscription?.product_id];
-        }
-        return undefined;
-    });
+    const product = useSelector(getSubscriptionProduct);
+    const subscription = useSelector(getCloudSubscription);
 
     if (!subscription || !product) {
         return null;
     }
 
-    const showSeatsAndSubscriptionDates = false;
     const isPaidTier = Boolean(subscription?.is_paid_tier === 'true');
-
-    const planPricing = getPlanPricing(isPaidTier, product);
 
     return (
         <div className='PlanDetails'>
             {planDetailsTopElements(userCount, isPaidTier, isFreeTrial, subscriptionPlan)}
-            {planPricing}
-            {showSeatsAndSubscriptionDates && seatsAndSubscriptionDates(locale, userCount, subscription.seats, new Date(subscription.start_at), new Date(subscription.end_at))}
+            <PlanPricing
+                isPaidTier={isPaidTier}
+                product={product}
+            />
             <div className='PlanDetails__teamAndChannelCount'>
                 <FormattedMessage
                     id='admin.billing.subscription.planDetails.features.unlimitedTeamsAndChannels'
