+++ components/admin_console/billing/plan_details/plan_details.scss
@@ -25,39 +25,12 @@
     font-size: 14px;
     font-weight: 600;
     line-height: 20px;
-
-    &.withinLimit {
-        color: var(--sys-sidebar-text-active-border);
-    }
-
-    &.overLimit {
-        color: var(--sys-error-text);
-    }
-}
-
-.PlanDetails__plan {
-    display: flex;
-    align-items: flex-end;
-    margin-top: 16px;
-    color: var(--sys-center-channel-color);
 }
 
 .PlanDetails__planName {
     font-size: 32px;
     font-weight: 600;
     line-height: 32px;
-}
-
-.PlanDetails_paidTier__planName {
-    color: rgba(var(--sys-center-channel-color-rgb), 0.56);
-    font-size: 12px;
-    font-style: normal;
-    font-weight: normal;
-    line-height: 16px;
-
-    a {
-        margin-left: 5px;
-    }
 }
 
 .PlanDetails__teamAndChannelCount {
@@ -88,25 +61,6 @@
     a::before { content: "\00a0"; }
 }
 
-.PlanDetails__feature {
-    margin-top: 12px;
-    color: var(--sys-center-channel-color);
-    line-height: 16px;
-
-    i {
-        font-size: 14.4px;
-
-        &::before {
-            margin: 0;
-        }
-    }
-
-    span {
-        margin-left: 8px;
-        font-size: 12px;
-    }
-}
-
 .PlanDetails__currentPlan {
     display: flex;
     align-items: center;
@@ -130,44 +84,6 @@
     }
 }
 
-.PlanDetails__seatsAndSubscriptionDates {
-    margin-top: 24px;
-    color: var(--sys-center-channel-color);
-    font-size: 14px;
-    line-height: 20px;
-}
-
-.PlanDetails__seats {
-    display: flex;
-    font-weight: 600;
-}
-
-.PlanDetails__seats-registered {
-    margin-left: 4px;
-    color: rgba(var(--sys-center-channel-color-rgb), 0.56);
-
-    &.overLimit {
-        color: var(--sys-error-text);
-    }
-
-    i {
-        margin-left: 4px;
-        font-size: 14.4px;
-        line-height: 17px;
-
-        &::before {
-            margin: 0;
-        }
-    }
-}
-
-.PlanDetails__subscriptionDate {
-    margin-top: 8px;
-}
-
-#BillingSubscriptions__seatOverageTooltip {
-    margin-left: -5px;
-}
 
 .TrialBadge {
     padding: 0 4px;
