+++ i18n/en.json
@@ -299,7 +299,6 @@
   "admin.billing.subscription.paymentFailed": "Payment failed. Please try again or contact support.",
   "admin.billing.subscription.paymentVerificationFailed": "Sorry, the payment verification failed",
   "admin.billing.subscription.planDetails.currentPlan": "(Current Plan)",
-  "admin.billing.subscription.planDetails.endDate": "End Date: ",
   "admin.billing.subscription.planDetails.features.10GBstoragePerUser": "10 GB storage per user",
   "admin.billing.subscription.planDetails.features.99uptime": "99.0% uptime",
   "admin.billing.subscription.planDetails.features.advanceTeamPermission": "Advanced team permissions",
@@ -330,16 +329,11 @@
   "admin.billing.subscription.planDetails.features.unlimittedUsersAndMessagingHistory": "Unlimited users & message history",
   "admin.billing.subscription.planDetails.flatFeePerMonth": "/month (Unlimited Users) ",
   "admin.billing.subscription.planDetails.howBillingWorks": "See how billing works",
-  "admin.billing.subscription.planDetails.numberOfSeats": "{numberOfSeats} seats",
-  "admin.billing.subscription.planDetails.numberOfSeatsRegistered": "({userCount} currently registered)",
   "admin.billing.subscription.planDetails.perUserPerMonth": "/user/month.",
   "admin.billing.subscription.planDetails.productName.cloudEnterprise": "Cloud Enterprise",
   "admin.billing.subscription.planDetails.productName.cloudProfessional": "Cloud Professional",
   "admin.billing.subscription.planDetails.productName.cloudStarter": "Cloud Starter",
   "admin.billing.subscription.planDetails.productName.mmCloud": "Mattermost Cloud",
-  "admin.billing.subscription.planDetails.prolongedOverages": "Prolonged overages may result in additional charges.",
-  "admin.billing.subscription.planDetails.seatCountOverages": "Seat count overages",
-  "admin.billing.subscription.planDetails.startDate": "Start Date: ",
   "admin.billing.subscription.planDetails.tiers.free": "Free",
   "admin.billing.subscription.planDetails.userCount": "{userCount} users",
   "admin.billing.subscription.privateCloudCard.cloudEnterprise.description": "At Mattermost, we work with you and your team to meet your needs throughout the product. If you are looking for an annual discount, please reach out to our sales team.",
